add support for the encls/enclu/enclv instructions (#5281)
Add support for the encls, enclu and enclv instructions as described by chapter 41 "SGX Instruction References" in volume 3D of the Intel 64 and IA-32 Architectures Software Developer's Manual. These are used in supervisor mode, user mode and VMM mode respectively.

These instructions are somewhat similar to cpuid, in that they have leaf functions that can be specified through the eax register. However, unlike cpuid, the ebx, ecx and edx can both be used as inputs and outputs depending on the specific leaf function.

While Intel SGX is usable on a select set of Intel CPUs, it is usable both in 32-bit and 64-bit mode on those CPUs.

The encls, enclu, enclv instructions are two-byte opcodes with the ModRM.reg field encoding the exact opcode. Since all four registers can be used as both implicit destination as well implicit source registers, this instruction takes extra operands, i.e. mrm and xop.

This PR is as follows:

 - Implement encls, enclu and enclv in the decoder for x86.
 - Add INSTR_CREATE macros for encls, enclu and enclv.
 - Add test cases for encls, enclu and enclv.
 - Fix comment for comment for cpuid describing that it uses 2 implicit sources and 4 implicit destinations instead to match the 4 implicit source/destination of encls/enclu/enclv.